requirements.txt
flask==3.0.0
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
python-dotenv==1.0.0
flask-login==0.6.3
werkzeug==3.0.1
requests==2.31.0
selenium==4.15.0
webdriver-manager==4.0.1
pandas==2.0.3
gunicorn==21.2.0
